remove types file
